truth-engine/requirements.txt
psycopg2-binary
sqlalchemy
pandas
numpy
alpaca-py
python-dotenv